Some file structure reformating, adding some syntax suger for the models, and starting implementation of better structure of code
