+++ LoremIpsum-SocialApp/MotelDatas/Album.swift
@@ -1,17 +1,16 @@
 //
-//  Comment.swift
+//  Album.swift
 //  LoremIpsum-SocialApp
 //
 //  Created by Lukasz Stachnik on 10/11/2020.
 //  Copyright © 2020 Warss. All rights reserved.
 //
-
 import Foundation
 
-struct Comment : Codable{
-    var postId: Int
+typealias Albums = [Album]
+
+struct Album : Codable{
+    var userId: Int
     var id: Int
-    var name: String
-    var email: String
-    var body: String
+    var title: String
 }
