+++ LoremIpsum-SocialApp/MotelDatas/Comment.swift
@@ -1,14 +1,19 @@
 //
-//  Album.swift
+//  Comment.swift
 //  LoremIpsum-SocialApp
 //
 //  Created by Lukasz Stachnik on 10/11/2020.
 //  Copyright © 2020 Warss. All rights reserved.
 //
+
 import Foundation
 
-struct Album : Codable{
-    var userId: Int
+typealias Comments = [Comment]
+
+struct Comment : Codable{
+    var postId: Int
     var id: Int
-    var title: String
+    var name: String
+    var email: String
+    var body: String
 }
